# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Crypto Trading Agent — Phase 1 observation API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Token {
  chain: string;
  address: string;
  symbol: string;
  name: string;
  /** @nullable */
  first_seen_at?: string | null;
  /** @nullable */
  latest_price_usd?: number | null;
  /** @nullable */
  latest_source?: string | null;
  /** @nullable */
  latest_collected_at?: string | null;
}

export interface PriceSnapshot {
  source: string;
  price_usd: number;
  /** @nullable */
  volume_24h_usd?: number | null;
  /** @nullable */
  liquidity_usd?: number | null;
  collected_at: string;
}

export interface HolderConcentration {
  /** @nullable */
  top_holder_pct?: number | null;
  /** @nullable */
  top_5_pct?: number | null;
  /** @nullable */
  top_10_pct?: number | null;
  /** @nullable */
  holder_count_checked?: number | null;
  note: string;
}

export interface SafetyAssessment {
  address: string;
  chain: string;
  mint_authority_active: boolean;
  freeze_authority_active: boolean;
  permanent_delegate_active: boolean;
  transfer_hook_active: boolean;
  /** @nullable */
  transfer_hook_program?: string | null;
  /** @nullable */
  transfer_fee_pct?: number | null;
  risk_score: number;
  flags: string[];
  checked_only: boolean;
  holder_concentration?: HolderConcentration | null;
  note: string;
}

export interface TechnicalReading {
  chain: string;
  address: string;
  data_points_used: number;
  /** @nullable */
  sma_20?: number | null;
  /** @nullable */
  ema_20?: number | null;
  /** @nullable */
  rsi_14?: number | null;
  /** @nullable */
  macd?: number | null;
  /** @nullable */
  macd_signal?: number | null;
  /** @nullable */
  macd_histogram?: number | null;
  /** @nullable */
  momentum_10?: number | null;
  /** @nullable */
  volatility_20?: number | null;
  note: string;
}

export type TradeRecommendationAction = typeof TradeRecommendationAction[keyof typeof TradeRecommendationAction];


export const TradeRecommendationAction = {
  buy: 'buy',
  sell: 'sell',
  hold: 'hold',
  ignore: 'ignore',
} as const;

export interface TradeRecommendation {
  chain: string;
  address: string;
  action: TradeRecommendationAction;
  /**
     * @minimum 0
     * @maximum 100
     */
  confidence: number;
  why_interesting: string;
  why_now: string;
  invalidation: string;
  downside: string;
  upside: string;
  missing_data: string[];
  model_used: string;
  actionable: boolean;
}

export interface SocialMention {
  platform: string;
  title: string;
  /** @nullable */
  url?: string | null;
  /** @nullable */
  score?: number | null;
  /** @nullable */
  num_comments?: number | null;
  created_at: string;
}

export type ListTokensParams = {
limit?: number;
};

export type GetTokenHistoryParams = {
limit?: number;
};

export type ListSocialMentionsParams = {
/**
 * @nullable
 */
token_symbol?: string | null;
limit?: number;
};