# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Crypto Trading Agent — Phase 1 observation API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: tokens
    description: Token market data
  - name: social
    description: Social mentions
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /tokens:
    get:
      operationId: listTokens
      tags: [tokens]
      summary: List tracked tokens with their latest price
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of tracked tokens
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Token"

  /tokens/{chain}/{address}/history:
    get:
      operationId: getTokenHistory
      tags: [tokens]
      summary: Price history for a token
      parameters:
        - name: chain
          in: path
          required: true
          schema:
            type: string
        - name: address
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 200
      responses:
        "200":
          description: Price history snapshots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PriceSnapshot"
        "404":
          description: Token not tracked

  /tokens/{chain}/{address}/safety:
    get:
      operationId: getTokenSafety
      tags: [tokens]
      summary: On-chain safety check and risk score
      parameters:
        - name: chain
          in: path
          required: true
          schema:
            type: string
        - name: address
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Safety assessment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SafetyAssessment"
        "404":
          description: Token not found on-chain

  /tokens/{chain}/{address}/technical:
    get:
      operationId: getTokenTechnical
      tags: [tokens]
      summary: Technical analysis indicators (SMA/EMA/RSI/MACD)
      parameters:
        - name: chain
          in: path
          required: true
          schema:
            type: string
        - name: address
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Technical analysis reading
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TechnicalReading"
        "404":
          description: Token not tracked

  /tokens/{chain}/{address}/recommendation:
    get:
      operationId: getTokenRecommendation
      tags: [tokens]
      summary: AI-generated trade recommendation (Claude)
      parameters:
        - name: chain
          in: path
          required: true
          schema:
            type: string
        - name: address
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Trade recommendation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeRecommendation"
        "404":
          description: Token not tracked
        "503":
          description: AI service unavailable

  /social/mentions:
    get:
      operationId: listSocialMentions
      tags: [social]
      summary: Recent social mentions from Reddit
      parameters:
        - name: token_symbol
          in: query
          schema:
            type: ["string", "null"]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Social mentions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SocialMention"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Token:
      type: object
      properties:
        chain:
          type: string
        address:
          type: string
        symbol:
          type: string
        name:
          type: string
        first_seen_at:
          type: ["string", "null"]
        latest_price_usd:
          type: ["number", "null"]
        latest_source:
          type: ["string", "null"]
        latest_collected_at:
          type: ["string", "null"]
      required:
        - chain
        - address
        - symbol
        - name

    PriceSnapshot:
      type: object
      properties:
        source:
          type: string
        price_usd:
          type: number
        volume_24h_usd:
          type: ["number", "null"]
        liquidity_usd:
          type: ["number", "null"]
        collected_at:
          type: string
      required:
        - source
        - price_usd
        - collected_at

    HolderConcentration:
      type: object
      properties:
        top_holder_pct:
          type: ["number", "null"]
        top_5_pct:
          type: ["number", "null"]
        top_10_pct:
          type: ["number", "null"]
        holder_count_checked:
          type: ["number", "null"]
        note:
          type: string
      required:
        - note

    SafetyAssessment:
      type: object
      properties:
        address:
          type: string
        chain:
          type: string
        mint_authority_active:
          type: boolean
        freeze_authority_active:
          type: boolean
        permanent_delegate_active:
          type: boolean
        transfer_hook_active:
          type: boolean
        transfer_hook_program:
          type: ["string", "null"]
        transfer_fee_pct:
          type: ["number", "null"]
        risk_score:
          type: number
        flags:
          type: array
          items:
            type: string
        checked_only:
          type: boolean
        holder_concentration:
          oneOf:
            - $ref: "#/components/schemas/HolderConcentration"
            - type: "null"
        note:
          type: string
      required:
        - address
        - chain
        - mint_authority_active
        - freeze_authority_active
        - permanent_delegate_active
        - transfer_hook_active
        - risk_score
        - flags
        - checked_only
        - note

    TechnicalReading:
      type: object
      properties:
        chain:
          type: string
        address:
          type: string
        data_points_used:
          type: number
        sma_20:
          type: ["number", "null"]
        ema_20:
          type: ["number", "null"]
        rsi_14:
          type: ["number", "null"]
        macd:
          type: ["number", "null"]
        macd_signal:
          type: ["number", "null"]
        macd_histogram:
          type: ["number", "null"]
        momentum_10:
          type: ["number", "null"]
        volatility_20:
          type: ["number", "null"]
        note:
          type: string
      required:
        - chain
        - address
        - data_points_used
        - note

    TradeRecommendation:
      type: object
      properties:
        chain:
          type: string
        address:
          type: string
        action:
          type: string
          enum: [buy, sell, hold, ignore]
        confidence:
          type: number
          minimum: 0
          maximum: 100
        why_interesting:
          type: string
        why_now:
          type: string
        invalidation:
          type: string
        downside:
          type: string
        upside:
          type: string
        missing_data:
          type: array
          items:
            type: string
        model_used:
          type: string
        actionable:
          type: boolean
      required:
        - chain
        - address
        - action
        - confidence
        - why_interesting
        - why_now
        - invalidation
        - downside
        - upside
        - missing_data
        - model_used
        - actionable

    SocialMention:
      type: object
      properties:
        platform:
          type: string
        title:
          type: string
        url:
          type: ["string", "null"]
        score:
          type: ["number", "null"]
        num_comments:
          type: ["number", "null"]
        created_at:
          type: string
      required:
        - platform
        - title
        - created_at